+++ docs/history.rst
@@ -4,6 +4,7 @@ Release History
 
 `Next Release`_
 ---------------
+- Added :class:`sprockets_status.handlers.StatusHandler`
 
 
 .. _Next Release: https://github.com/sprockets/sprockets-status/compare/0.0.0...HEAD